Add Validation to Images

+++ app/Http/Controllers/Admin/ImageController.php
@@ -7,44 +7,57 @@ use Illuminate\Http\Request;
 use App\Models\Image;
 use App\Models\Apartment;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Auth;
 
 class ImageController extends Controller
 {
 
     public function index(Apartment $apartment)
     {
-        return view('admin.apartments.images.index', ['apartment' => $apartment]);
+        if ($apartment->user_id === Auth::id()) {
+            return view('admin.apartments.images.index', ['apartment' => $apartment]);
+        } else {
+            abort(403);
+        }
     }
 
 
     public function setMain(Apartment $apartment, Image $image)
     {
-        $main_image = Image::where('apartment_id', $apartment->id)->where('is_main', '1')->first();
-        $main_image->update([
-            'is_main' => 0
-        ]);
+        if ($apartment->user_id === Auth::id()) {
+            $main_image = Image::where('apartment_id', $apartment->id)->where('is_main', '1')->first();
+            $main_image->update([
+                'is_main' => 0
+            ]);
 
-        $image->update([
-            'is_main' => 1
-        ]);
+            $image->update([
+                'is_main' => 1
+            ]);
 
-        return to_route('admin.apartments.images.index', $apartment)->with('message', 'Immagine principale sostituita con successo.');
+            return to_route('admin.apartments.images.index', $apartment)->with('message', 'Immagine principale sostituita con successo.');
+        } else {
+            abort(403);
+        }
     }
 
     public function destroy(Apartment $apartment, Image $image)
     {
-        $was_main = $image->is_main;
+        if ($apartment->user_id === Auth::id()) {
+            $was_main = $image->is_main;
 
-        Storage::delete($image->path);
-        $image->delete();
+            Storage::delete($image->path);
+            $image->delete();
 
-        if ($was_main) {
-            $new_main_image = Image::where('apartment_id', $apartment->id)->first();
-            $new_main_image->update([
-                'is_main' => 1
-            ]);
+            if ($was_main) {
+                $new_main_image = Image::where('apartment_id', $apartment->id)->first();
+                $new_main_image->update([
+                    'is_main' => 1
+                ]);
+            }
+
+            return to_route('admin.apartments.images.index', $apartment)->with('message', 'Immagine eliminata con successo.');
+        } else {
+            abort(403);
         }
-
-        return to_route('admin.apartments.images.index', $apartment)->with('message', 'Immagine eliminata con successo.');
     }
 }
